fix(#4222): address fourth roborev round — full WHERE-coverage guard, LIMIT-exemption parity, keyspace-guard false positive, generation width
Roborev review round 4 (job 4320) reported 4 findings; all fixed:

- [High] The OR/NOT-only WHERE guard missed every OTHER silently-dropped
  comparison shape (!=, NotIn, LIKE, NotLike, NotBetween, IS [NOT] NULL,
  Regex, non-literal RHS) — column_comparison_to_predicate returns None
  for all of these even inside a perfectly pushable AND position.
  Replaced with count_pushable_comparison_leaves: walks the WHERE tree
  counting Comparison leaves in a pushable position (None if OR/NOT
  appears anywhere), and compares that count against
  plan.sstable_predicates.len() — catching both classes, including
  `pk = 1 AND val != 'x'` (a non-empty predicate list that a bare
  emptiness check would miss).
- [Medium] The full-scan producer's LIMIT-exemption disagreed with the
  point-key producer's PointCollector: under an explicit LIMIT, the
  full-scan path still applied the row-count valve unconditionally,
  while the point path (and the final check) exempted it. Both
  producers now agree: an explicit LIMIT exempts max_result_rows,
  leaving only the byte budget.
- [Medium] resolve_raw_view_readers's keyspace-fallback guard fired on
  ANY !fully_qualified_match, including the benign case of a
  schema-loaded-but-no-SSTables table (readers.is_empty()) — which that
  flag cannot distinguish from a genuine bare-name-fallback collision.
  Now gated on !readers.is_empty() too, so an empty result falls through
  honestly instead of a misleading keyspace-mismatch error.
- [Low] generation is now bigint (was int, saturating on any u64 above
  i32::MAX and collapsing two distinct generations to the same
  fabricated value) — widened the column type and its Value::BigInt
  emission; updated all test helpers/assertions accordingly.

Added regression tests: where_clause_with_not_equal_fails_closed
(the narrower gap the OR/NOT-only guard missed).

+++ cqlite-core/src/query/select_executor/raw_view/columns.rs
@@ -178,7 +178,13 @@ pub(in crate::query::select_executor) fn raw_view_columns(
     )?;
 
     push(&mut columns, "sstable".to_string(), "text")?;
-    push(&mut columns, "generation".to_string(), "int")?;
+    // `bigint`, not `int` (roborev finding, issue #4222): `SSTableReader::generation`
+    // is a `u64`; narrowing it to `i32` would SATURATE (and thus collapse
+    // two distinct generations to the same fabricated value) for any real
+    // corpus whose generation identifiers exceed `i32::MAX` — a
+    // no-heuristics violation in a view whose whole purpose is authoritative
+    // per-generation source identity.
+    push(&mut columns, "generation".to_string(), "bigint")?;
     push(&mut columns, "format".to_string(), "text")?;
     push(&mut columns, "position".to_string(), "bigint")?;
